pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "sift-stack-py>=0.4.2",
    "streamlit>=1.43.2",
]
